Restructure into runtime scripts and base sections of astropt
