+++ scripts/scaling/galaxy.png
