+++ scripts/scaling/hexbin_xkcd.png
