+++ scripts/scaling/scaling_xkcd.png
